requirements.txt
streamlit>=1.16.0
openai>=1.0.0
faiss-cpu>=1.7.3
numpy>=1.21.0
langchain>=0.3.0,<0.4
langchain-core>=0.3.0,<0.4
langchain-openai>=0.2.0,<0.3
langchain-google-genai>=2.0.0,<3
langchain-community>=0.3.0,<0.4
langchain-text-splitters>=0.3.0,<0.4
python-dotenv
pandas>=1.4.0
rank-bm25>=0.2.2
PyMuPDF>=1.23.0
pydantic>=2.0.0
redis>=4.0.0
streamlit-authenticator>=0.4.2
google-auth>=2.26.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-cloud-storage>=2.10.0
cryptography>=42.0.0
argon2-cffi>=23.1.0
tenacity>=8.0.0
structlog>=23.0.0